+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# develop
+
+# [1.1.0](https://github.com/bklang/sequella/compare/v1.0.0...v1.1.0) - [2014-01-14](https://rubygems.org/gems/sequella/versions/1.1.0)
+  * Feature: Rake tasks to clean and reset the database
+  * Feature: Automatically use JDBC on JRuby
+  * Cleanup: No more autoloading (or dependency on activesupport)
+
 # [1.0.0](https://github.com/bklang/sequella/tags/v1.0.0) - [2013-05-20](https://rubygems.org/gems/sequella/versions/1.0.0)
 
 * Initial release